+++ examples/changelog/docs/changelog.md
@@ -1,23 +1,1 @@
 # Changelog
-
-## New Classes
-
-These classes are new.
-
-### AccountService
-
-This is a new class that does foo and bar and references Baz .
-
-## New Interfaces
-
-These interfaces are new.
-
-### IAnotherExample
-
-### IExemplificable
-
-## New Enums
-
-These enums are new.
-
-### PossibleValues
