+++ src/core/renderables/documentables.ts
@@ -1,7 +1,7 @@
 import { CustomTag, RenderableDocumentation, RenderableContent } from './types';
 import { Describable, Documentable, GetRenderableContentByTypeName } from './types';
-import { replaceInlineReferences } from './inline';
-import { isEmptyLine } from './type-utils';
+import { replaceInlineReferences } from '../markdown/adapters/inline';
+import { isEmptyLine } from '../markdown/adapters/type-utils';
 
 export function adaptDescribable(
   describable: Describable,
